partial feat: add support methods
#find_moves(origin) method provides an array with all the legal moves a knight
might make from where it currently is. This shall be used to create  a data tree
where each possible move is a node of the tree, and each node points to several
other possible moves from there using again the #find_moves(origin) in a
recursive way.

The #path_search method is the search algorithm that should traverse the tree
in BFS to find the SHORTEST path to the desired destination. It needs to
implement a FIFO queue that stores each move's subsequent possible moves in
a long list to be later iterated through in the next recursion "level" until the
desired destiny node is reached.

#knight_moves should just print the result without messing the recursion. If the
result can be printed with the recursion this method might not be necessary.