feat: find number of moves to reach destiny
The #path_search method can find the number of moves it takes to reach desired
destination, but since the possible paths of each step are being reduced it can
not keep track of the path taken to reach the destination.

The 'origin' included as last element of the array returned by #find_moves shall
provide the reference to this path on recursive calls, but to do so the queue
passed as argument to the next call can't be reduced: the recursive calls should
somehow be able to keep track of the path being taken to destiny.