feat: print path taken to destination
The path taken to each position is included in that position's list of next level
possible moves to keep track of the individual path being taken on succesive
recursive calls until destination if found. The addition of this path reference
had to be done on the #knight_moves definition instead of the #find_moves
because only #knight_moves could provide the reference to the full path
without a callback.